Add mutex in plot node

+++ wb_controller/src/plot_node.cpp
@@ -18,9 +18,18 @@
 #include <wb_controller/TerrainEstimation.h>
 #include <wb_controller/FootHolds.h>
 #include <rviz_visual_tools/rviz_visual_tools.h>
+#include <mutex>
+
+/*
+ * BLUE: DESIRED
+ * GREEN: ACTUAL
+ *
+ */
 
 namespace wb_controller
 {
+
+static std::mutex _mtx;
 
 template <class msg_t>
 class Visualizer
@@ -54,7 +63,7 @@ protected:
       pose_.translation().y() = position.y;
       pose_.translation().z() = position.z;
 
-      visual_tools_->publishCone(pose_,M_PI);
+      visual_tools_->publishCone(pose_,M_PI,rviz_visual_tools::colors::TRANSLUCENT_LIGHT,0.1); // FIXME use the relation between mu and angle
       visual_tools_->trigger();
   }
 
@@ -77,7 +86,7 @@ protected:
   }
 
   // ARROW
-  void createArrow(const geometry_msgs::Vector3& vector, const geometry_msgs::Vector3& position,  rviz_visual_tools::colors color)
+  void createArrow(const geometry_msgs::Vector3& vector, const geometry_msgs::Vector3& origin,  rviz_visual_tools::colors color, double scale = 500.0)
   {
       vector_(0) = vector.x;
       vector_(1) = vector.y;
@@ -87,13 +96,31 @@ protected:
       //find rotation matrix to align 1 0  0 to force direction
       R_ = Eigen::Quaterniond().setFromTwoVectors(Eigen::Vector3d::UnitX(),vector_/norm_).toRotationMatrix();
       pose_.linear() = R_;
-      pose_.translation().x() = position.x;
-      pose_.translation().y() = position.y;
-      pose_.translation().z() = position.z;
-      visual_tools_->publishArrow(pose_, color, rviz_visual_tools::LARGE, norm_/500.0);
+      pose_.translation().x() = origin.x;
+      pose_.translation().y() = origin.y;
+      pose_.translation().z() = origin.z;
+      visual_tools_->publishArrow(pose_, color, rviz_visual_tools::LARGE, norm_/scale);
       visual_tools_->trigger();
   }
 
+  void createArrow(const geometry_msgs::Vector3& vector, const geometry_msgs::Point& origin,  rviz_visual_tools::colors color, double scale = 500.0)
+  {
+      vector_(0) = vector.x;
+      vector_(1) = vector.y;
+      vector_(2) = vector.z;
+      norm_ = vector_.norm()+0.00001;
+
+      //find rotation matrix to align 1 0  0 to force direction
+      R_ = Eigen::Quaterniond().setFromTwoVectors(Eigen::Vector3d::UnitX(),vector_/norm_).toRotationMatrix();
+      pose_.linear() = R_;
+      pose_.translation().x() = origin.x;
+      pose_.translation().y() = origin.y;
+      pose_.translation().z() = origin.z;
+      visual_tools_->publishArrow(pose_, color, rviz_visual_tools::LARGE, norm_/scale);
+      visual_tools_->trigger();
+  }
+
+  // SPHERE
   void createSphere(const geometry_msgs::Point& origin, rviz_visual_tools::colors color)
   {
     pose_.translation().x() = origin.x;
@@ -120,8 +147,8 @@ protected:
   Eigen::Matrix3d R_;
   ros::Subscriber subscriber_;
   rviz_visual_tools::RvizVisualToolsPtr visual_tools_;
-
 };
+
 
 class FrictionConesVisualizer : public Visualizer<wb_controller::FrictionCones>
 {
@@ -141,11 +168,12 @@ protected:
 
   virtual void callback(const wb_controller::FrictionCones &msg) override
   {
-    if(cnt_++%decimate_==0)
+    if(cnt_++%decimate_==0 && _mtx.try_lock())
     {
         visual_tools_->deleteAllMarkers();
         for(unsigned int i=0; i<msg.foot_positions.size();i++)
            createCone(msg.cone_axis[i],msg.foot_positions[i]);
+        _mtx.unlock();
     }
   }
 };
@@ -167,10 +195,11 @@ public:
 protected:
   virtual void callback(const wb_controller::TerrainEstimation &msg) override
   {
-    if(cnt_++%decimate_==0)
+    if(cnt_++%decimate_==0 && _mtx.try_lock())
     {
         visual_tools_->deleteAllMarkers();
         createPlane(msg.terrain_normal,msg.central_point);
+        _mtx.unlock();
     }
   }
 };
@@ -191,7 +220,7 @@ protected:
 
   virtual void callback(const wb_controller::ContactForces &msg) override
   {
-    if(cnt_++%decimate_==0)
+    if(cnt_++%decimate_==0 && _mtx.try_lock())
     {
         visual_tools_->deleteAllMarkers();
         //Display an arrow along the x-axis of a pose.
@@ -199,6 +228,7 @@ protected:
         {
            createArrow(msg.des_contact_forces[i].force,msg.contact_positions[i],rviz_visual_tools::BLUE);
            createArrow(msg.contact_forces[i].force,msg.contact_positions[i],rviz_visual_tools::GREEN);
+           _mtx.unlock();
         }
     }
   }
@@ -220,10 +250,12 @@ protected:
 
   virtual void callback(const wb_controller::CartesianTask& msg) override
   {
-    if(cnt_++%decimate_==0)
+    if(cnt_++%decimate_==0 && _mtx.try_lock())
     {
         visual_tools_->deleteAllMarkers();
         createSphere(msg.pose_actual.position,rviz_visual_tools::GREEN);
+        createArrow(msg.twist_reference.linear,msg.pose_actual.position,rviz_visual_tools::BLUE,1.0);
+        _mtx.unlock();
     }
   }
 };
@@ -243,13 +275,14 @@ class FootHoldsVisualizer : public Visualizer<wb_controller::FootHolds>
 protected:
   virtual void callback(const wb_controller::FootHolds& msg) override
   {
-    if(cnt_++%decimate_==0)
+    if(cnt_++%decimate_==0 && _mtx.try_lock())
     {
         visual_tools_->deleteAllMarkers();
         for(unsigned int i=0;i<msg.name.size();i++)
         {
           //createSphere(msg.desired_foothold[i],rviz_visual_tools::BLUE);
           createSphere(msg.virtual_foothold[i],rviz_visual_tools::RED);
+          _mtx.unlock();
         }
     }
   }
@@ -267,9 +300,10 @@ int main(int argc, char **argv)
   ros::NodeHandle node("wb_controller");
 
   wb_controller::ContactForcesVisualizer cfv(node,"contact_forces");
-  wb_controller::CoMVisualizer comv(node,"com");
+  wb_controller::CoMVisualizer comv(node,"CoM");
   wb_controller::FootHoldsVisualizer fhv(node,"foot_holds");
   wb_controller::TerrainEstimationVisualizer tev(node,"terrain_estimation");
+  wb_controller::FrictionConesVisualizer fcv(node,"friction_cones");
 
   ros::spin();
